# Conflicts:
#	pom.xml
#	src/main/java/com/ntv/ntvcons_backend/constants/EntityType.java
#	src/main/java/com/ntv/ntvcons_backend/controllers/ProjectBlueprintController.java
#	src/main/java/com/ntv/ntvcons_backend/dtos/blueprint/BlueprintDTO.java
#	src/main/java/com/ntv/ntvcons_backend/dtos/project/ProjectDTO.java
#	src/main/java/com/ntv/ntvcons_backend/entities/BaseEntity.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Blueprint.java
#	src/main/java/com/ntv/ntvcons_backend/entities/EntityWrapper.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ExternalFile.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ExternalFileEntityWrapperPairing.java
#	src/main/java/com/ntv/ntvcons_backend/entities/FileType.java
#	src/main/java/com/ntv/ntvcons_backend/entities/FileTypeModels/ShowFileTypeModel.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Location.java
#	src/main/java/com/ntv/ntvcons_backend/entities/LocationModels/ShowLocationModel.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Post.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Project.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ProjectBlueprintModels/ShowProjectBlueprintModel.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ProjectManager.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ProjectWorker.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Report.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ReportDetail.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ReportModels/ShowReportModel.java
#	src/main/java/com/ntv/ntvcons_backend/entities/ReportType.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Request.java
#	src/main/java/com/ntv/ntvcons_backend/entities/RequestDetail.java
#	src/main/java/com/ntv/ntvcons_backend/entities/RequestType.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Role.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Task.java
#	src/main/java/com/ntv/ntvcons_backend/entities/TaskAssignment.java
#	src/main/java/com/ntv/ntvcons_backend/entities/TaskReport.java
#	src/main/java/com/ntv/ntvcons_backend/entities/User.java
#	src/main/java/com/ntv/ntvcons_backend/entities/Worker.java
#	src/main/java/com/ntv/ntvcons_backend/entities/externalFileModels/ShowExternalFileModel.java
#	src/main/java/com/ntv/ntvcons_backend/entities/projectModels/ShowProjectModel.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/EntityWrapperRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ExternalFileEntityWrapperPairingRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ExternalFileRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/FileTypeRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/LocationRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/PostRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ProjectBlueprintRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ProjectManagerRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ProjectRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ProjectWorkerRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ReportDetailRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ReportRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/ReportTypeRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/RequestDetailRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/RequestRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/RequestTypeRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/RoleRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/TaskAssignmentRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/TaskReportRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/TaskRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/UserRepository.java
#	src/main/java/com/ntv/ntvcons_backend/repositories/WorkerRepository.java
#	src/main/java/com/ntv/ntvcons_backend/services/fileType/FileTypeService.java
#	src/main/java/com/ntv/ntvcons_backend/services/location/LocationService.java
#	src/main/java/com/ntv/ntvcons_backend/services/location/LocationServiceImpl.java
#	src/main/java/com/ntv/ntvcons_backend/services/project/ProjectServiceImpl.java
#	src/main/java/com/ntv/ntvcons_backend/services/projectBlueprint/ProjectBlueprintService.java
#	src/main/java/com/ntv/ntvcons_backend/services/projectBlueprint/ProjectBlueprintServiceImpl.java
#	src/main/java/com/ntv/ntvcons_backend/services/report/ReportServiceImpl.java
#	src/main/java/com/ntv/ntvcons_backend/services/reportDetail/ReportDetailServiceImpl.java
#	src/main/java/com/ntv/ntvcons_backend/services/role/RoleService.java
#	src/main/java/com/ntv/ntvcons_backend/services/role/RoleServiceImpl.java

+++ src/main/java/com/ntv/ntvcons_backend/entities/FileTypeModels/ShowFileTypeModel.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ShowFileTypeModel {
-    private int fileTypeId;
+    private long fileTypeId;
     private String fileTypeName, fileTypeDesc, fileTypeExtension;
 
 }
